split binary search from sortedIndexWith
