+++ src/sortedIndexBy.ts
@@ -1,5 +1,5 @@
 import { purry } from './purry';
-import { sortedIndexWithImplementation } from './sortedIndexWith';
+import { binarySearchCutoffIndex } from './_binarySearchCutoffIndex';
 
 /**
  * Find the insertion position (index) of an item in an array with items sorted
@@ -117,7 +117,7 @@ function sortedIndexByImplementation<T>(
   valueFunction: (item: T, index?: number) => NonNullable<unknown>
 ): number {
   const value = valueFunction(item);
-  return sortedIndexWithImplementation(
+  return binarySearchCutoffIndex(
     array,
     (pivot, index) => valueFunction(pivot, index) < value
   );
