+++ src/sortedIndexWith.test.ts
@@ -1,176 +1,20 @@
+/**
+ * This utility function simply delegates the call to `binarySearchCutoffIndex`,
+ * so we test it mainly via the tests for that function.
+ */
 import { sortedIndexWith } from './sortedIndexWith';
 
-describe('runtime correctness', () => {
-  test('empty array', () => {
-    expect(sortedIndexWith([], () => true)).toBe(0);
+describe('sanity', () => {
+  test('regular', () => {
+    expect(sortedIndexWith([0, 1, 2, 3, 4, 5, 6, 7, 8, 9], x => x < 5)).toBe(5);
   });
 
-  test('single value, always false', () => {
-    expect(sortedIndexWith([1], () => false)).toBe(0);
-  });
-
-  test('single value, always true', () => {
-    expect(sortedIndexWith([1], () => true)).toBe(1);
-  });
-
-  test('multiple values, always false', () => {
-    expect(sortedIndexWith([1, 2, 3, 4, 5], () => false)).toBe(0);
-  });
-
-  test('multiple values, always true', () => {
-    expect(sortedIndexWith([1, 2, 3, 4, 5], () => true)).toBe(5);
-  });
-
-  test('fancy stuff', () => {
+  test('indexed', () => {
     expect(
-      sortedIndexWith(
-        ['a', 'ab', 'abc', 'abcd', 'abcde'],
-        ({ length }) => length < 3
+      sortedIndexWith.indexed(
+        ['a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j'],
+        (_, index) => index < 5
       )
-    ).toBe(2);
+    ).toBe(5);
   });
 });
-
-describe('binary search correctness via indexed', () => {
-  test('after 0 items', () => {
-    expect(
-      indicesSeen(
-        [
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-        ],
-        (_, index) => index < 0
-      )
-    ).toEqual([8, 4, 2, 1, 0]);
-  });
-
-  test('after 4 items', () => {
-    expect(
-      indicesSeen(
-        [
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-        ],
-        (_, index) => index < 4
-      )
-    ).toEqual([8, 4, 2, 3]);
-  });
-
-  test('after 8 items', () => {
-    expect(
-      indicesSeen(
-        [
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-        ],
-        (_, index) => index < 8
-      )
-    ).toEqual([8, 4, 6, 7]);
-  });
-
-  test('after 12 items', () => {
-    expect(
-      indicesSeen(
-        [
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-        ],
-        (_, index) => index < 12
-      )
-    ).toEqual([8, 12, 10, 11]);
-  });
-
-  test('after 20 items', () => {
-    expect(
-      indicesSeen(
-        [
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-          undefined,
-        ],
-        (_, index) => index < 20
-      )
-    ).toEqual([8, 12, 14, 15]);
-  });
-});
-
-function indicesSeen(
-  items: ReadonlyArray<unknown>,
-  predicate: (item: unknown, index: number) => boolean
-): ReadonlyArray<number> {
-  const indicesSeen: Array<number> = [];
-  sortedIndexWith.indexed(items, (pivot, index) => {
-    indicesSeen.push(index);
-    return predicate(pivot, index);
-  });
-  return indicesSeen;
-}
