+++ src/sortedLastIndex.ts
@@ -1,5 +1,5 @@
 import { purry } from './purry';
-import { sortedIndexWithImplementation } from './sortedIndexWith';
+import { binarySearchCutoffIndex } from './_binarySearchCutoffIndex';
 
 /**
  * Find the insertion position (index) of an item in an array with items sorted
@@ -57,7 +57,7 @@ const sortedLastIndexImplementation = <T>(
   array: ReadonlyArray<T>,
   item: T
 ): number =>
-  sortedIndexWithImplementation(
+  binarySearchCutoffIndex(
     array,
     // The only difference between the regular implementation and the "last"
     // variation is that we consider the pivot with equality too, so that we
